execution/vm/benchmark: move TestSnailtracerPathsAgree next to the benchmark
